# Impose limits on decoded integers via HPACK