Addresses: https://github.com/aspnet/KestrelHttpServer/issues/2922. Setting the limit to Int32.MaxValue for now since that's also the limit for request header field size.