@JunTaoLuo 2018-09-25
Good idea

On Tue, Sep 25, 2018 at 9:40 AM Chris Ross <notifications@github.com<mailto:notifications@github.com>> wrote:

@Tratcher commented on this pull request.

________________________________

In test/Kestrel.Core.Tests/IntegerDecoderTests.cs<https://github.com/aspnet/KestrelHttpServer/pull/2953#discussion_r220265192>:

> @@ -34,6 +34,21 @@ public void IntegerDecode(int i, int prefixLength, byte[] octets)
             Assert.Equal(i, decoder.Value);
         }


Can you add a round trip test? for 0 to max int, encode, decode, and verify?

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/aspnet/KestrelHttpServer/pull/2953#pullrequestreview-158633012>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AB768ztaKSjcSEeH97FRCzpRarHk8oh5ks5uelxqgaJpZM4W3ncx>.