@JunTaoLuo 2018-09-25
Might want to take another look since I added a bit of optimization.

Before this PR:
```
                  Method |      Mean |     Error |    StdDev |          Op/s | Scaled | ScaledSD | Allocated |
------------------------ |----------:|----------:|----------:|--------------:|-------:|---------:|----------:|
 DecodeSingleByteInteger |  1.126 ns | 0.0273 ns | 0.0255 ns | 888,439,368.7 |   1.00 |     0.00 |       0 B |
  DecodeMultiByteInteger | 15.122 ns | 0.2951 ns | 0.3030 ns |  66,130,977.7 |  13.44 |     0.39 |       0 B |
```
After this PR:
```
                  Method |       Mean |     Error |    StdDev |            Op/s | Scaled | ScaledSD | Allocated |
------------------------ |-----------:|----------:|----------:|----------------:|-------:|---------:|----------:|
 DecodeSingleByteInteger |  0.8466 ns | 0.0057 ns | 0.0053 ns | 1,181,147,635.1 |   1.00 |     0.00 |       0 B |
  DecodeMultiByteInteger | 15.7103 ns | 0.0728 ns | 0.0568 ns |    63,652,630.6 |  18.56 |     0.13 |       0 B |
```

Note the multibyte scenario didn't improve since we limit checks which mostly balanced the gains.